+++ lib/brando/blueprint.ex
@@ -153,7 +153,6 @@ defmodule Brando.Blueprint do
     Module.put_attribute(__CALLER__.module, :gettext_module, gettext_module)
 
     Module.register_attribute(__CALLER__.module, :ctx, accumulate: false)
-    Module.register_attribute(__CALLER__.module, :meta_fields, accumulate: true)
     Module.register_attribute(__CALLER__.module, :traits, accumulate: true)
     Module.register_attribute(__CALLER__.module, :attrs, accumulate: true)
     Module.register_attribute(__CALLER__.module, :relations, accumulate: true)
@@ -185,6 +184,7 @@ defmodule Brando.Blueprint do
 
       use Ecto.Schema
       use Brando.Blueprint.JSONLD
+      use Brando.Blueprint.Meta
 
       require PolymorphicEmbed
       import Ecto.Changeset
@@ -197,7 +197,6 @@ defmodule Brando.Blueprint do
       import unquote(__MODULE__).Identifier
       import unquote(__MODULE__).Listings
       import unquote(__MODULE__).Listings.Components
-      import unquote(__MODULE__).Meta
       import unquote(__MODULE__).Naming
       import unquote(__MODULE__).Relations
       import unquote(__MODULE__).Trait