+++ lib/brando/blueprint/meta/meta_field.ex
@@ -1,0 +1,4 @@
+defmodule Brando.Blueprint.Meta.MetaField do
+  defstruct targets: [],
+            value_fn: nil
+end