+++ lib/brando/blueprint/meta/meta_schema.ex
@@ -1,0 +1,4 @@
+defmodule Brando.Blueprint.Meta.MetaSchema do
+  defstruct __identifier__: nil,
+            fields: []
+end